Fix mic button behavior
This change introduces a custom View subclass for the microphone
tab button, handling tap and swipe events internally instead of
using `OnSwipe` and `OnClick` on the motion layout in which the
button is embedded. The latter handlers conflicted both with
each other and with click handlers registered on the button itself.

This also removes the need for the tray fragment itself to listen
to MotionLayout transition events.